+++ src/routing/verb/http-verb-helpers.ts
@@ -1,6 +1,6 @@
 import type { HttpMiddleware, HttpRequest } from '@/Http';
-import type { RouteDefinition } from '@/routing/route/definition/route-definition.type';
 import { Route } from '@/routing/route/route';
+import type { RouteDefinition } from '@/routing/route/route-definition.type';
 import type { HttpMethod } from '@/routing/verb/http-method.type';
 import type { Request } from 'koa';
